modify type of BlockID's blknum as i32

+++ src/file/block_id.rs
@@ -3,11 +3,11 @@ use core::fmt;
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
 pub struct BlockId {
     filename: String,
-    blknum: u64,
+    blknum: i32,
 }
 
 impl BlockId {
-    pub fn new(filename: &str, blknum: u64) -> Self {
+    pub fn new(filename: &str, blknum: i32) -> Self {
         Self {
             filename: filename.to_string(),
             blknum,
@@ -18,7 +18,7 @@ impl BlockId {
         self.filename.clone()
     }
 
-    pub fn number(&self) -> u64 {
+    pub fn number(&self) -> i32 {
         self.blknum
     }
 }
